exploration/requirements.txt
alpaca-py==0.34.0
numpy>=1.26.0
pandas==2.2.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
ta-lib==0.5.1
scikit-learn>=1.3.0
scikit-optimize==0.9.0
matplotlib==3.8.3
seaborn==0.13.2
